Add files via upload
OpenCV projects - It contains a list of different programs and operations. Edge detection is done using Sobel operator, Key point detection using Difference of Gaussian. The SIFT features is used for Panorama - Image stitching . The Hough transform is used to detect lines, the morphological transforms are used to remove noise from an image.